aklkl

+++ lib/app/modules/simulator/charts/model-chart.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 
 class ChartSubscribers {
   final String mes;
-  final int geracao;
+  final double geracao;
   final charts.Color barColor;
 
   ChartSubscribers({@required this.mes, @required this.geracao, @required this.barColor});
